requirements.txt
Django==1.11.29
djangorestframework==3.9.3
factory-boy==2.12.0
